自动清理

+++ xsy-common/src/main/java/com/xsy/base/util/Clean.java
@@ -1,6 +1,7 @@
 package com.xsy.base.util;
 
 import lombok.Getter;
+import lombok.Setter;
 import lombok.Synchronized;
 import lombok.extern.slf4j.Slf4j;
 
@@ -9,6 +10,8 @@ import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
+import java.util.Collections;
+import java.util.List;
 
 /**
  * @author Q1sj
@@ -18,17 +21,38 @@ import java.time.ZoneId;
 @Getter
 public class Clean {
     private final static Object LOCK = new Object();
+    /**
+     * 清理目录
+     */
     private final File path;
+    /**
+     * 清理几天前的数据
+     */
     private final int cleanDayAgo;
-    private final long reservedSpaceSize;
+    /**
+     * 允许最大使用率 默认85
+     * 此参数优先级大于{@link #cleanDayAgo}
+     * 使用率大于此值会继续删除
+     */
+    @Setter
+    private int allowMaxUtilizationRate;
+    /**
+     * 要删除的文件扩展名
+     */
+    private final List<String> extensionList;
 
-    public Clean(String path, int cleanDayAgo, long reservedSpaceSize) {
-        if (cleanDayAgo <= 0 && reservedSpaceSize <= 0) {
+    public Clean(String path, int cleanDayAgo) {
+        this(path, cleanDayAgo, Collections.emptyList());
+    }
+
+    public Clean(String path, int cleanDayAgo, List<String> extensionList) {
+        if (cleanDayAgo <= 0) {
             throw new IllegalArgumentException();
         }
+        this.allowMaxUtilizationRate = 85;
         this.path = new File(path);
         this.cleanDayAgo = cleanDayAgo;
-        this.reservedSpaceSize = reservedSpaceSize;
+        this.extensionList = Collections.unmodifiableList(extensionList);
     }
 
     @Synchronized("LOCK")
@@ -38,27 +62,34 @@ public class Clean {
             return;
         }
         int cleanDayAgo = this.cleanDayAgo;
-        log.info("{} 删除{}天前数据", path, cleanDayAgo);
+        log.info("{} 删除{}天前数据 扩展名:{}", path, cleanDayAgo, extensionList.isEmpty() ? "*" : extensionList);
         clean(path, cleanDayAgo);
         long usableSpace = path.getUsableSpace();
-        log.info("{} 需要预留空间:{} 可用空间:{}", path, FileUtils.byteCountToDisplaySize(reservedSpaceSize), FileUtils.byteCountToDisplaySize(usableSpace));
-        while (path.getUsableSpace() < reservedSpaceSize && cleanDayAgo > 0) {
+        int utilizationRate = utilizationRate();
+        log.info("{} 可用空间:{} 使用率:{}% 允许最大使用率:{}%", path, FileUtils.byteCountToDisplaySize(usableSpace), utilizationRate, allowMaxUtilizationRate);
+        while (allowMaxUtilizationRate < utilizationRate && cleanDayAgo > 0) {
             if (!this.path.exists()) {
                 log.warn("目录{}不存在", this.path.getAbsolutePath());
                 return;
             }
             cleanDayAgo--;
-            log.info("{} 预留空间不足 删除{}天前数据", path, cleanDayAgo);
+            log.info("{} 删除{}天前数据 扩展名:{}", path, cleanDayAgo, extensionList.isEmpty() ? "*" : extensionList);
             clean(path, cleanDayAgo);
-            log.info("{} 需要预留空间:{} 可用空间:{}", path, FileUtils.byteCountToDisplaySize(reservedSpaceSize), FileUtils.byteCountToDisplaySize(usableSpace));
+            utilizationRate = utilizationRate();
+            log.info("{} 可用空间:{} 使用率:{}% 允许最大使用率:{}%", path, FileUtils.byteCountToDisplaySize(usableSpace), utilizationRate, allowMaxUtilizationRate);
         }
+    }
+
+    private int utilizationRate() {
+        double rate = 100D * (path.getTotalSpace() - path.getUsableSpace()) / path.getTotalSpace();
+        return (int) rate;
     }
 
     private void clean(File file, int cleanDayAgo) {
         if (file.isDirectory()) {
             File[] files = file.listFiles();
-            if (files == null) {
-                log.info("删除空目录 {}", file.getAbsolutePath());
+            if (files == null || files.length == 0) {
+                log.warn("删除空目录 {}", file.getAbsolutePath());
                 FileUtils.deleteQuietly(file);
                 return;
             }
@@ -70,8 +101,10 @@ public class Clean {
         LocalDateTime localDateTime = LocalDateTime.ofInstant(Instant.ofEpochMilli(file.lastModified()), ZoneId.systemDefault());
         LocalDateTime now = LocalDate.now().atStartOfDay();
         if (now.plusDays(-cleanDayAgo).isAfter(localDateTime)) {
-            log.info("lastModified:{} delete {}", localDateTime, file);
-            FileUtils.deleteQuietly(file);
+            if (extensionList.isEmpty() || extensionList.contains(FileUtils.getExtension(file.getName()))) {
+                log.warn("delete lastModified:{} {}", localDateTime, file);
+                FileUtils.deleteQuietly(file);
+            }
         }
     }
 }
